Stock Handler
This assignment's technical details is the Random.NextDouble(), which instead of generating - numbers and +, it's just generating random numbers. Which saves me from adding more code and keeping it simple. Some of the clever aspects that were added to this assignment are the custom-formatted output, which makes it easier to read.