Project II
Some technical details that are in my project 2. This project is about 3 card games, and I coded them in one package, allowing the players to play all 3. In the Apples and Oranges part of the game, the tuple return is a good technical detail I need to mention; this part allows two cards to be drawn at the same time, which benefits a lot, helps me not add too much or extra variables, and is less confusing for the player.